reorganize stage stuff renamed DirectorRegister to ActManager and moved everthing else to that namespace.
