+++ card/mod/core/set/all/actify.rb
@@ -19,6 +19,41 @@ def abort status, msg="action canceled"
   raise Card::Error::Abort.new(status, msg)
 end
 
+def act opts={}
+  if ActManager.act_card
+    add_to_act opts
+    main_act_block = false
+  else
+    start_new_act opts
+    main_act_block = true
+  end
+  yield
+ensure
+  ActManager.clear if main_act_block
+end
+
+def start_new_act opts
+  ActManager.clear
+  self.director = nil
+  ActManager.act_card = self
+  if opts && opts[:success]
+    Env[:success] = Env::Success.new(cardname, Env.params[:success])
+  end
+end
+
+def add_to_act opts
+  # if only_storage_phase is true the card is already part of the act
+  return if ActManager.act_card == self || only_storage_phase
+  director.reset_stage
+  if opts && opts[:trash]
+    @action = :delete
+  else
+    identify_action
+  end
+  director.update_card self
+  self.only_storage_phase = true
+end
+
 module ClassMethods
   def create! opts
     card = Card.new opts
