+++ card/spec/lib/card/stage_director_spec.rb
@@ -1,4 +1,4 @@
-describe Card::StageDirector do
+describe Card::ActManager::StageDirector do
   describe "abortion" do
     let(:create_card) { Card.create name: "a card" }
     let(:create_card_with_subcard) do
